[light] Cap effects per light and store the active effect index in 16 bits
Nothing limited how many effects a light could have, and the active index was a uint32_t. Effects cost tens of bytes of RAM each, so a light with more than 65535 of them is not a real configuration; validation now rejects one, and LightState stores the index as a uint16_t. Public signatures stay uint32_t, and the saved state keeps its uint32_t field so upgrading does not reset every light's restored state.

+++ esphome/components/light/light_state.cpp
@@ -339,7 +339,7 @@ void LightState::start_effect_(uint32_t effect_index) {
   if (effect_index == 0)
     return;
 
-  this->active_effect_index_ = effect_index;
+  this->active_effect_index_ = static_cast<uint16_t>(effect_index);  // validate_() range-checked it
   auto *effect = this->get_active_effect_();
   effect->start_internal();
   // Enable loop while effect is active
